Add Integer and Decimal input types
