+++ views/inputs/integer.php
@@ -13,6 +13,6 @@ defined( 'ABSPATH' ) || exit;
 $input = $this->input;
 
 $input['type']  = 'text';
-$input['class'] = sprintf( '%s %s', $input['class'] ?? '', 'gla-input-number' );
+$input['class'] = sprintf( '%s %s', $input['class'] ?? '', 'gla-input-integer' );
 
 woocommerce_wp_text_input( $input );
